requirements.txt
Flask==2.0.0
Flask-Compress==1.9.0
mako
markdown>=3.0
nbformat==5.1.3
nbconvert==6.0.7
parso==0.8.2
pdoc3==0.9.2